Web-App/requirements.txt
tkinterdnd2==0.4.3
torch==2.6.0+cpu
torchvision==0.21.0+cpu
facenet-pytorch==2.5.0
opencv-python==4.11.0.86
Pillow==11.0.0
numpy==2.2.4
tqdm==4.67.1
transformers==4.51.0
huggingface-hub==0.30.1
accelerate==1.2.1
tokenizers==0.21.1
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
requests==2.32.3
streamlit==1.40.1
